+++ backend/sparrow_tests/scripts/dump-test-database.py
@@ -12,18 +12,22 @@ from sparrow.logs import console_handler, get_logger
 from sparrow_tests.fixtures import basic_project
 import logging
 
-log = get_logger(level=logging.DEBUG, handler=console_handler)
+log = get_logger(level=logging.WARNING, handler=console_handler)
+
 
 conn = "postgresql://postgres@db:5432/sparrow_test_1"
 with redirect_stdout(stderr), testing_database(conn) as engine:
-    a1 = App(__name__)
-    a1.database.initialize()
     # Re-initialize app
-    app = App(__name__ + "1")
+    app = App(__name__)
+    app.database.initialize()
     app.load()
     app.load_phase_2()
 
-    project = app.database.load_data("project", basic_project)
+    logging.getLogger("sparrow.interface.schema").setLevel(logging.ERROR)
+
+    project = app.database.load_data(
+        "project", basic_project, session=app.database.session
+    )
 
     dz_data = import_dz_test_data()
     session = app.database.load_data("session", dz_data)
@@ -41,9 +45,11 @@ with redirect_stdout(stderr), testing_database(conn) as engine:
     project.session_collection = [session]
 
     app.database.session.add(sample)
+    app.database.session.add(project)
+    app.database.session.commit()
 
     # Prepare for dump
-    app.database.session.execute("TRUNCATE TABLE spatial_ref_sys;")
+    app.database.engine.execute("TRUNCATE TABLE spatial_ref_sys")
 
     dbargs = connection_args(engine)
-    run("pg_dump", "-Fc", "-Z9", dbargs, engine.url.database, stdout=stdout)
+    run("pg_dump", "-Fc", dbargs, engine.url.database, stdout=stdout)